move TCPConn
